network info

+++ lib/features/camera/data/repositories/camera_repository_impl.dart
@@ -3,6 +3,7 @@ import 'package:dartz/dartz.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:my_camera_app_demo/cores/exceptions/exception.dart';
 import 'package:my_camera_app_demo/cores/failures/failure.dart';
+import 'package:my_camera_app_demo/cores/network/network_info.dart';
 import 'package:my_camera_app_demo/features/camera/data/datasources/local_camera_datasource.dart';
 import 'package:my_camera_app_demo/features/camera/data/datasources/remote_camera_datasource.dart';
 import 'package:my_camera_app_demo/features/camera/data/models/picture_model.dart';
@@ -11,10 +12,12 @@ import 'package:my_camera_app_demo/features/camera/domain/repositories/camera_re
 class CameraRepositoryImpl implements CameraRepository {
   final LocalCameraDatasource localCameraDatasource;
   final RemoteCameraDatasource remoteCameraDatasource;
+  final NetworkInfo networkInfo;
 
   CameraRepositoryImpl({
     @required this.localCameraDatasource,
     @required this.remoteCameraDatasource,
+    @required this.networkInfo,
   });
 
   @override
@@ -32,15 +35,18 @@ class CameraRepositoryImpl implements CameraRepository {
         lastModifyTime: DateTime.now().toUtc(),
       );
       try {
-        await remoteCameraDatasource.sendPicture(jwt, model);
+        if (await networkInfo.isConnected) {
+          await remoteCameraDatasource.sendPicture(jwt, model);
+          await localCameraDatasource.deleteFile(path);
+          return Right(Unit);
+        } else
+          throw RemoteSavePictureException();
       } on RemoteSavePictureException {
         final result = await localCameraDatasource.savePicture(model);
+        await localCameraDatasource.deleteFile(path);
         if (result == 0) return Left(SavePictureFailure());
         return Right(Unit);
       }
-
-      await localCameraDatasource.deleteFile(path);
-      return Right(Unit);
     } on GetFileException {
       return Left(GetFileFailure());
     } on DeleteFileException {
